+++ src/node/miner.h
@@ -154,11 +154,11 @@ struct CompareModifiedEntry {
             // high gas limit but a low gas price which has a child with a low gas limit but a high gas price
             // Without this condition that transaction chain would get priority in being included into the block.
             // The two next checks are to see if all our ancestors have been added.
-            if(a.nSizeWithAncestors == a.iter->GetTxSize() && b.nSizeWithAncestors != b.iter->GetTxSize()) {
+            if((int64_t) a.nSizeWithAncestors == a.iter->GetTxSize() && (int64_t) b.nSizeWithAncestors != b.iter->GetTxSize()) {
                 return true;
             }
 
-            if(b.nSizeWithAncestors == b.iter->GetTxSize() && a.nSizeWithAncestors != a.iter->GetTxSize()) {
+            if((int64_t) b.nSizeWithAncestors == b.iter->GetTxSize() && (int64_t) a.nSizeWithAncestors != a.iter->GetTxSize()) {
                 return false;
             }
